+++ packages/default-field-editors/README.md
@@ -7,5 +7,5 @@ npm install @contentful/field-editor-field
 This package contains a React `Field` component which is used to render different field editors based on the widget ID.
 
 ```js
-import { Field } from '@contentful/field-editor-field';
+import { Field } from 'packages/default-field-editors';
 ```
