+++ packages/default-field-editors/src/index.tsx
@@ -1,3 +1,3 @@
 export { Field } from './Field';
-export type { WidgetType } from './Field';
 export { FieldWrapper } from './FieldWrapper';
+export type { WidgetType } from './types';
