+++ packages/default-field-editors/src/types.ts
@@ -1,0 +1,26 @@
+export type WidgetType =
+  | 'multipleLine'
+  | 'boolean'
+  | 'objectEditor'
+  | 'datePicker'
+  | 'locationEditor'
+  | 'checkbox'
+  | 'listInput'
+  | 'rating'
+  | 'radio'
+  | 'tagEditor'
+  | 'numberEditor'
+  | 'urlEditor'
+  | 'slugEditor'
+  | 'singleLine'
+  | 'dropdown'
+  | 'entryLinkEditor'
+  | 'entryCardEditor'
+  | 'entryLinksEditor'
+  | 'entryCardsEditor'
+  | 'assetLinkEditor'
+  | 'assetLinksEditor'
+  | 'assetGalleryEditor'
+  | 'richTextEditor'
+  | 'markdown'
+  | string;
